// src/app/models/purchase-invoice-detail-model.ts
export class PurchaseInvoiceDetailModel {
  primaryKey?: string;
  invoicePrimaryKey: string;
  orderPrimaryKey: string;
  orderDetailPrimaryKey: string;
  productPrimaryKey: string;
  listPrice?: number; // price on list
  price?: number; // price of employee typed
  defaultPrice?: number; // converted price of list price
  discount1?: number;
  defaultDiscount1?: number;
  discount2?: number;
  defaultDiscount2?: number;
  quantity?: number;
  taxRate?: number;
  insertDate?: number;
  totalPrice?: number; // iskonto1 ve iskont2 dusmus hali.
  totalPriceWithTax?: number; // totalPrice + kdv
  unitPrimaryKey: string;
  unitValue: number;
}